# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "kiaranote is simple notion-like service",
        "title": "kiaranote",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Ryo",
            "url": "https://github.com/bbconfhq/kiaranote",
            "email": "gwanryo@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/bbconfhq/kiaranote/blob/main/LICENSE"
        },
        "version": "0.1"
    },
    "basePath": "/v1",
    "paths": {
        "/auth/login": {
            "get": {
                "description": "Request login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.response"
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "get": {
                "description": "Request logout",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Logout",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.response"
                        }
                    }
                }
            }
        },
        "/healthcheck": {
            "get": {
                "description": "Check health",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "healthcheck"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "integer"
                }
            }
        }
    }
}